Update travis to use Chef 12.1.0.
The current version of the build-essential cookbook depends on
compat_resource which munges multi-line package calls with string
arguments into a single line mulitpackage resource call with an
array argument containing all the packages. Unfortuntatly,
multipackage was not introduced until Chef 12.1.0, so we have to bump
the Chef version we test against to support this feature.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>