Refactor erlang_install to support latest kerl.
There were a number of breaking changes in kerl over the past year. The
biggest change is that the install directory must either be empty or
not exist. As result, this is a breaking change in the cookbook as we
can no longer install at top level directories such as `/usr/local`. A
helper method was added to raise an error when the prefix path does not
meet this requirement, before attempting build to prevent a long delay
before identifying the issue.

Additionally, kerl returns a non-zero status when a build request is
made for an existing build. New logic was added to make the the build
idempotent.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>